--- test/fixtures/big5.json
@@ -1,260 +1,0 @@
-{
-    "bytes": [
-        60,
-        72,
-        84,
-        77,
-        76,
-        62,
-        13,
-        10,
-        60,
-        72,
-        69,
-        65,
-        68,
-        62,
-        32,
-        32,
-        32,
-        32,
-        13,
-        10,
-        9,
-        60,
-        84,
-        73,
-        84,
-        76,
-        69,
-        62,
-        32,
-        109,
-        101,
-        116,
-        97,
-        32,
-        188,
-        208,
-        197,
-        210,
-        170,
-        186,
-        168,
-        207,
-        165,
-        206,
-        161,
-        71,
-        164,
-        164,
-        164,
-        229,
-        186,
-        244,
-        173,
-        182,
-        32,
-        60,
-        47,
-        84,
-        73,
-        84,
-        76,
-        69,
-        62,
-        13,
-        10,
-        9,
-        60,
-        109,
-        101,
-        116,
-        97,
-        32,
-        72,
-        84,
-        84,
-        80,
-        45,
-        69,
-        81,
-        85,
-        73,
-        86,
-        61,
-        34,
-        67,
-        111,
-        110,
-        116,
-        101,
-        110,
-        116,
-        45,
-        84,
-        121,
-        112,
-        101,
-        34,
-        32,
-        67,
-        79,
-        78,
-        84,
-        69,
-        78,
-        84,
-        61,
-        34,
-        116,
-        101,
-        120,
-        116,
-        47,
-        104,
-        116,
-        109,
-        108,
-        59,
-        32,
-        99,
-        104,
-        97,
-        114,
-        115,
-        101,
-        116,
-        61,
-        98,
-        105,
-        103,
-        53,
-        34,
-        62,
-        13,
-        10,
-        60,
-        47,
-        72,
-        69,
-        65,
-        68,
-        62,
-        13,
-        10,
-        60,
-        66,
-        79,
-        68,
-        89,
-        62,
-        13,
-        10,
-        13,
-        10,
-        179,
-        111,
-        172,
-        79,
-        164,
-        64,
-        173,
-        211,
-        193,
-        99,
-        197,
-        233,
-        164,
-        164,
-        164,
-        229,
-        186,
-        244,
-        173,
-        182,
-        161,
-        73,
-        60,
-        98,
-        114,
-        62,
-        13,
-        10,
-        40,
-        84,
-        104,
-        105,
-        115,
-        32,
-        112,
-        97,
-        103,
-        101,
-        32,
-        117,
-        115,
-        101,
-        115,
-        32,
-        98,
-        105,
-        103,
-        53,
-        32,
-        99,
-        104,
-        97,
-        114,
-        97,
-        99,
-        116,
-        101,
-        114,
-        32,
-        115,
-        101,
-        116,
-        46,
-        41,
-        60,
-        98,
-        114,
-        62,
-        13,
-        10,
-        99,
-        104,
-        97,
-        114,
-        115,
-        101,
-        116,
-        61,
-        98,
-        105,
-        103,
-        53,
-        13,
-        10,
-        13,
-        10,
-        60,
-        47,
-        66,
-        79,
-        68,
-        89,
-        62,
-        13,
-        10,
-        60,
-        47,
-        72,
-        84,
-        77,
-        76,
-        62
-    ],
-    "string": "<HTML>\r\n<HEAD>    \r\n\t<TITLE> meta 標籤的使用：中文網頁 </TITLE>\r\n\t<meta HTTP-EQUIV=\"Content-Type\" CONTENT=\"text/html; charset=big5\">\r\n</HEAD>\r\n<BODY>\r\n\r\n這是一個繁體中文網頁！<br>\r\n(This page uses big5 character set.)<br>\r\ncharset=big5\r\n\r\n</BODY>\r\n</HTML>"
-}
